+++ frontend/src/components/navigation/Navbar.js
@@ -1,6 +1,7 @@
 import { Nav, Navbar, Container } from "react-bootstrap";
 import { useState, useEffect } from "react";
-import '../styles/App.css'
+import '../../styles/App.css'
+
 export default function NavBar() {
     const [username, setUsername] = useState('');
