+++ frontend/src/pages/Home.js
@@ -1,5 +1,5 @@
 import React, { useEffect, useState } from 'react';
-import NavBar from '../components/Navbar';
+import NavBar from '../components/navigation/Navbar';
 import '../styles/App.css';
 import GameApp from '../components/game/GameApp';
